streamlit-portfolio/requirements.txt
streamlit==1.31.1
pandas==2.2.0
plotly==5.18.0
pillow==10.2.0
requests==2.31.0
streamlit-lottie==0.0.5
streamlit-option-menu==0.3.12 